SkillNotFoundExc added

+++ src/main/java/lt/swedbank/services/skill/SkillService.java
@@ -4,9 +4,12 @@ package lt.swedbank.services.skill;
 import lt.swedbank.beans.entity.Skill;
 import lt.swedbank.beans.request.AddSkillRequest;
 import lt.swedbank.beans.request.RemoveSkillRequest;
+import lt.swedbank.exceptions.skill.SkillNotFaoundException;
 import lt.swedbank.repositories.SkillRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class SkillService implements ISkillService {
@@ -29,8 +32,11 @@ public class SkillService implements ISkillService {
     }
 
     @Override
-    public Skill removeSkill(Long userID, RemoveSkillRequest removeSkillRequest)
+    public Skill removeSkill(Long userID, RemoveSkillRequest removeSkillRequest) throws SkillNotFaoundException
     {
+        if (!Optional.ofNullable(skillRepository.findByTitleAndUserID(removeSkillRequest.getTitle(), userID)).isPresent()) {
+            throw new SkillNotFaoundException();
+        }
         Skill skill = skillRepository.findByTitleAndUserID(removeSkillRequest.getTitle(), userID);
         skillRepository.delete(skill);
         return skill;
